+++ src/Compiler/IR/LIR/HIRToLIRLowerer.Lowering.Expressions.MemberAssignments.cs
@@ -13,7 +13,7 @@ public sealed partial class HIRToLIRLowerer
     private bool UsesStrictAssignmentSemantics()
         => _scope == null || Jroc.Utilities.ArgumentsObjectSemantics.IsStrictScope(_scope);
 
-    private bool TryLowerPropertyAssignmentTarget(HIRExpression objectExpr, string propertyName, TempVariable valueToStore, out TempVariable resultTempVar)
+    private bool TryLowerPropertyAssignmentTarget(HIRExpression objectExpr, string propertyName, TempVariable valueToStore, out TempVariable resultTempVar, bool resultUsed = true)
     {
         resultTempVar = default;
 
@@ -30,7 +30,7 @@ public sealed partial class HIRToLIRLowerer
                 IsPrivateField: false,
                 fieldValueToStore));
 
-            resultTempVar = fieldValueToStore;
+            resultTempVar = resultUsed ? fieldValueToStore : default;
             return true;
         }
 
@@ -39,10 +39,10 @@ public sealed partial class HIRToLIRLowerer
             return false;
         }
 
-        return TryLowerPropertyAssignmentTarget(objTemp, propertyName, valueToStore, out resultTempVar);
+        return TryLowerPropertyAssignmentTarget(objTemp, propertyName, valueToStore, out resultTempVar, resultUsed);
     }
 
-    private bool TryLowerPropertyAssignmentTarget(TempVariable objectTemp, string propertyName, TempVariable valueToStore, out TempVariable resultTempVar)
+    private bool TryLowerPropertyAssignmentTarget(TempVariable objectTemp, string propertyName, TempVariable valueToStore, out TempVariable resultTempVar, bool resultUsed = true)
     {
         resultTempVar = default;
 
@@ -64,11 +64,11 @@ public sealed partial class HIRToLIRLowerer
         var setResult = CreateTempVariable();
         _methodBodyIR.Instructions.Add(new LIRSetItem(objectTemp, boxedKey, valueToStore, setResult, UsesStrictAssignmentSemantics()));
         DefineTempStorage(setResult, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
-        resultTempVar = setResult;
+        resultTempVar = resultUsed ? setResult : default;
         return true;
     }
 
-    private bool TryLowerIndexAssignmentTarget(HIRExpression objectExpr, HIRExpression indexExpr, TempVariable valueToStore, out TempVariable resultTempVar)
+    private bool TryLowerIndexAssignmentTarget(HIRExpression objectExpr, HIRExpression indexExpr, TempVariable valueToStore, out TempVariable resultTempVar, bool resultUsed = true)
     {
         resultTempVar = default;
 
@@ -88,7 +88,7 @@ public sealed partial class HIRToLIRLowerer
                 IsPrivateField: false,
                 fieldValueToStore));
 
-            resultTempVar = fieldValueToStore;
+            resultTempVar = resultUsed ? fieldValueToStore : default;
             return true;
         }
 
@@ -102,10 +102,10 @@ public sealed partial class HIRToLIRLowerer
             return false;
         }
 
-        return TryLowerIndexAssignmentTarget(objTemp, indexTemp, valueToStore, out resultTempVar);
+        return TryLowerIndexAssignmentTarget(objTemp, indexTemp, valueToStore, out resultTempVar, resultUsed);
     }
 
-    private bool TryLowerIndexAssignmentTarget(TempVariable objectTemp, TempVariable indexTemp, TempVariable valueToStore, out TempVariable resultTempVar)
+    private bool TryLowerIndexAssignmentTarget(TempVariable objectTemp, TempVariable indexTemp, TempVariable valueToStore, out TempVariable resultTempVar, bool resultUsed = true)
     {
         resultTempVar = default;
 
@@ -129,11 +129,11 @@ public sealed partial class HIRToLIRLowerer
         var setResult = CreateTempVariable();
         _methodBodyIR.Instructions.Add(new LIRSetItem(objectTemp, indexForSet, valueToStore, setResult, UsesStrictAssignmentSemantics()));
         DefineTempStorage(setResult, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
-        resultTempVar = setResult;
+        resultTempVar = resultUsed ? setResult : default;
         return true;
     }
 
-    private bool TryLowerPropertyAssignmentExpression(HIRPropertyAssignmentExpression assignExpr, out TempVariable resultTempVar)
+    private bool TryLowerPropertyAssignmentExpression(HIRPropertyAssignmentExpression assignExpr, out TempVariable resultTempVar, bool resultUsed = true)
     {
         resultTempVar = default;
 
@@ -160,7 +160,6 @@ public sealed partial class HIRToLIRLowerer
             _methodBodyIR.Instructions.Add(new LIRGetItem(objTemp, boxedKey, current));
             DefineTempStorage(current, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
 
-            resultTempVar = CreateTempVariable();
             var evalRightLabel = CreateLabel();
             var endLabel = CreateLabel();
 
@@ -172,7 +171,11 @@ public sealed partial class HIRToLIRLowerer
             DefineTempStorage(isJsNullTemp, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
             _methodBodyIR.Instructions.Add(new LIRBranchIfTrue(isJsNullTemp, evalRightLabel));
 
-            _methodBodyIR.Instructions.Add(new LIRCopyTemp(currentBoxed, resultTempVar));
+            if (resultUsed)
+            {
+                resultTempVar = CreateTempVariable();
+                _methodBodyIR.Instructions.Add(new LIRCopyTemp(currentBoxed, resultTempVar));
+            }
             _methodBodyIR.Instructions.Add(new LIRBranch(endLabel));
 
             _methodBodyIR.Instructions.Add(new LIRLabel(evalRightLabel));
@@ -187,11 +190,17 @@ public sealed partial class HIRToLIRLowerer
             var setResult = CreateTempVariable();
             _methodBodyIR.Instructions.Add(new LIRSetItem(objTemp, boxedKey, rhsValue, setResult, UsesStrictAssignmentSemantics()));
             DefineTempStorage(setResult, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
-            _methodBodyIR.Instructions.Add(new LIRCopyTemp(setResult, resultTempVar));
+            if (resultUsed)
+            {
+                _methodBodyIR.Instructions.Add(new LIRCopyTemp(setResult, resultTempVar));
+            }
 
             _methodBodyIR.Instructions.Add(new LIRLabel(endLabel));
             ClearNumericRefinementsAtLabel();
-            DefineTempStorage(resultTempVar, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
+            if (resultUsed)
+            {
+                DefineTempStorage(resultTempVar, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
+            }
             return true;
         }
 
@@ -209,7 +218,7 @@ public sealed partial class HIRToLIRLowerer
                     return false;
                 }
 
-                return TryLowerPropertyAssignmentTarget(assignExpr.Object, assignExpr.PropertyName, valueToStore, out resultTempVar);
+                return TryLowerPropertyAssignmentTarget(assignExpr.Object, assignExpr.PropertyName, valueToStore, out resultTempVar, resultUsed);
             }
 
             if (!TryLowerExpression(assignExpr.Object, out var objTemp))
@@ -222,7 +231,7 @@ public sealed partial class HIRToLIRLowerer
                 return false;
             }
 
-            return TryLowerPropertyAssignmentTarget(objTemp, assignExpr.PropertyName, valueToStore, out resultTempVar);
+            return TryLowerPropertyAssignmentTarget(objTemp, assignExpr.PropertyName, valueToStore, out resultTempVar, resultUsed);
         }
         else
         {
@@ -251,7 +260,7 @@ public sealed partial class HIRToLIRLowerer
             }
         }
 
-        return TryLowerPropertyAssignmentTarget(assignExpr.Object, assignExpr.PropertyName, valueToStore, out resultTempVar);
+        return TryLowerPropertyAssignmentTarget(assignExpr.Object, assignExpr.PropertyName, valueToStore, out resultTempVar, resultUsed);
     }
 
     /// <summary>
@@ -343,7 +352,7 @@ public sealed partial class HIRToLIRLowerer
         return true;
     }
 
-    private bool TryLowerIndexAssignmentExpression(HIRIndexAssignmentExpression assignExpr, out TempVariable resultTempVar)
+    private bool TryLowerIndexAssignmentExpression(HIRIndexAssignmentExpression assignExpr, out TempVariable resultTempVar, bool resultUsed = true)
     {
         resultTempVar = default;
 
@@ -377,7 +386,6 @@ public sealed partial class HIRToLIRLowerer
             _methodBodyIR.Instructions.Add(new LIRGetItem(objTemp, indexForGet, current));
             DefineTempStorage(current, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
 
-            resultTempVar = CreateTempVariable();
             var evalRightLabel = CreateLabel();
             var endLabel = CreateLabel();
 
@@ -389,7 +397,11 @@ public sealed partial class HIRToLIRLowerer
             DefineTempStorage(isJsNullTemp, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
             _methodBodyIR.Instructions.Add(new LIRBranchIfTrue(isJsNullTemp, evalRightLabel));
 
-            _methodBodyIR.Instructions.Add(new LIRCopyTemp(currentBoxed, resultTempVar));
+            if (resultUsed)
+            {
+                resultTempVar = CreateTempVariable();
+                _methodBodyIR.Instructions.Add(new LIRCopyTemp(currentBoxed, resultTempVar));
+            }
             _methodBodyIR.Instructions.Add(new LIRBranch(endLabel));
 
             _methodBodyIR.Instructions.Add(new LIRLabel(evalRightLabel));
@@ -410,11 +422,17 @@ public sealed partial class HIRToLIRLowerer
             var setResult = CreateTempVariable();
             _methodBodyIR.Instructions.Add(new LIRSetItem(objTemp, indexForSet, rhsValue, setResult, UsesStrictAssignmentSemantics()));
             DefineTempStorage(setResult, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
-            _methodBodyIR.Instructions.Add(new LIRCopyTemp(setResult, resultTempVar));
+            if (resultUsed)
+            {
+                _methodBodyIR.Instructions.Add(new LIRCopyTemp(setResult, resultTempVar));
+            }
 
             _methodBodyIR.Instructions.Add(new LIRLabel(endLabel));
             ClearNumericRefinementsAtLabel();
-            DefineTempStorage(resultTempVar, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
+            if (resultUsed)
+            {
+                DefineTempStorage(resultTempVar, new ValueStorage(ValueStorageKind.Reference, typeof(object)));
+            }
             return true;
         }
 
@@ -435,7 +453,7 @@ public sealed partial class HIRToLIRLowerer
                     return false;
                 }
 
-                return TryLowerIndexAssignmentTarget(assignExpr.Object, assignExpr.Index, valueToStore, out resultTempVar);
+                return TryLowerIndexAssignmentTarget(assignExpr.Object, assignExpr.Index, valueToStore, out resultTempVar, resultUsed);
             }
 
             if (!TryLowerExpression(assignExpr.Object, out var objTemp))
@@ -453,7 +471,7 @@ public sealed partial class HIRToLIRLowerer
                 return false;
             }
 
-            return TryLowerIndexAssignmentTarget(objTemp, indexTemp, valueToStore, out resultTempVar);
+            return TryLowerIndexAssignmentTarget(objTemp, indexTemp, valueToStore, out resultTempVar, resultUsed);
         }
         else
         {
@@ -497,7 +515,7 @@ public sealed partial class HIRToLIRLowerer
             }
         }
 
-        return TryLowerIndexAssignmentTarget(assignExpr.Object, assignExpr.Index, valueToStore, out resultTempVar);
+        return TryLowerIndexAssignmentTarget(assignExpr.Object, assignExpr.Index, valueToStore, out resultTempVar, resultUsed);
     }
 
     private bool TryLowerDestructuringAssignmentExpression(HIRDestructuringAssignmentExpression assignExpr, out TempVariable resultTempVar)